+++ doc/conf.py
@@ -3,8 +3,6 @@
 This configuration only affects single-package Sphinx documentation builds.
 """
 
-import os
-
 import lsst.ts.imsim  # noqa
 from documenteer.conf.pipelinespkg import *  # noqa
 
@@ -15,13 +13,8 @@ html_short_title = project
 doxylink = {}
 
 
-# Support the sphinx extension of plantuml
-extensions.append("sphinxcontrib.plantuml")  # noqa
-
-# Put the path to plantuml.jar
-plantuml_path = (
-    os.path.expanduser("~/plantuml.jar")
-    if os.getenv("PATH_PLANTUML") is None
-    else os.getenv("PATH_PLANTUML")
-)
-plantuml = f"java -jar {plantuml_path}"
+# Support the sphinx extension of mermaid
+extensions = [
+    "sphinxcontrib.mermaid",
+    "sphinx_automodapi.automodapi",
+]