+++ python/lsst/ts/imsim/closed_loop_task.py
@@ -616,9 +616,9 @@ class ClosedLoopTask:
             if cam_type == CamType.LsstCam:
                 default_pointer = "lsstCamDefaultPointer.yaml"
             elif cam_type == CamType.LsstFamCam:
-                default_pointer = "lsstFamCamDefaultPointer.yaml"
+                default_pointer = "lsstFamNoPertPointer.yaml"
             elif cam_type == CamType.ComCam:
-                default_pointer = "lsstComCamDefaultPointer.yaml"
+                default_pointer = "lsstComCamNoPertPointer.yaml"
             imsim_config_pointer_file = os.path.join(get_config_dir(), default_pointer)
 
         base_config_yaml = self.imsim_cmpt.assemble_config_yaml(
@@ -1423,7 +1423,8 @@ config.dataset_config.ref_dataset_name='ref_cat'
             help="""
                  Text file contains the star Id, ra, dec, and magnitude.
                  The default is to use the OPD field positions with boresight
-                 [ra, dec] = [0, 0].
+                 [ra, dec] = [0, 0]. The file header must include column names
+                 Id, Ra, Dec, Mag. For example, see tests/testData/sky/wfsStar.txt.
                  """,
         )